1.2.2:
Prior to this release there were separate checkboxes for "index, follow", "noindex" and "nofollow", even though checkboxes are nonexclusive, so it was possible to pick "index, follow" and "noindex" simultaneously — an invalid combination. Beginning with this release, there is no "index, follow" option because that is the default behavior of Google and never has to be explicitly chosen. If you want your page to be crawled and indexed normally, just leave the "noindex" and "nofollow" checkboxes alone.

However, for bc reasons, if you are already using this module you may note that "index, follow" is explicitly output on pages for which you haven't edited the settings yet. That's fine and will have no ill effects. The only case you might want to look into is anywhere you chose an invalid combination of options as described above.